+++ client/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 
 - Fixed timeline graph search links calling search-estimates endpoints for non-active tabs ([#1007](https://github.com/project-lux/lux-frontend/issues/1007)).
 - Fixed unused code and refactored naming convention of searchTag vs halLink ([#1004](https://github.com/project-lux/lux-frontend/issues/1004)).
+- Fixed the timeline list accessibility ([#865](https://github.com/project-lux/lux-frontend/issues/865)).
 
 ### Removed